+++ smallvm/process.cpp
@@ -6,6 +6,8 @@
 #include "runtime/parserengine.h"
 #include <QDebug>
 
+#include <cstdint>
+
 #define assert0(cond, toSignal) if(!(cond)) { signal(toSignal); }
 #define assert1(cond, toSignal, arg0) if(!(cond)) { signal(toSignal, arg0); }
 #define assert2(cond, toSignal, arg0, arg1) if(!(cond)) { signal(toSignal, arg0, arg1); }
@@ -453,7 +455,7 @@ IMPLEMENT_VM_INSTRUCTION(DoPushLocal)
 
 IMPLEMENT_VM_INSTRUCTION(DoPushLocalCached)
 {
-    Value **v = currentFrame()->fastLocals + (int) executedInstruction->Arg;
+    Value **v = currentFrame()->fastLocals + (intptr_t) executedInstruction->Arg;
     assert1(v, NoSuchVariable1, executedInstruction->SymRef);
     pushOperand(*v);
 }
